posts now in better defined file heirarchy
